backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pymongo==4.8.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
pydantic[email]==2.8.2
python-dotenv==1.0.1
scikit-learn==1.5.2
numpy==1.26.4
joblib==1.4.2
